test(usni): extract fleet parser helpers
